Let VeinBlock be an interface and have different implementations
This was attempted in the past but ultimately scrapped due to how BlockData worked at the time. However, now is a better time than ever to abstractify how VeinBlocks are handled due to the different ways to check if types are encapsulated. Comparison with VeinBlock instances that only require type definitions should be much faster now that BlockData is no longer involved in the computation.

I'm still not happy about "raw data", but this is a hacky work around until a more proper solution may be found or pull requested to Bukkit. Until such time, it will remain in datable vein blocks.