Add a categories.yml and let tool categories be entirely configurable
Categories may be dynamically defined and utilized in the config.yml blocklist. Each category may specify a list of tools that belong to said category with either just a material, or some additional meta such as name and lore. In addition to that, each category (and specific tool) may further override some of the global-scoped configuration options including MaxVeinSize, IncludesEdges, DisabledWorlds and RepairFriendlyVeinMiner.

This commit also fixes a couple bugs
- Fixed MaxVeinSize being 1 more than specified in the config
- Fixed non-durable items not capable of vein mining
- Fixed the hand category not being recognized (Fixes #29)

Known caveats:
- Permission nodes do not work for VeinMiner. They were temporarily removed and will be added back in a later commit