Don't integrate mcMMO-Classic
Registering this listener with mcMMO-Classic installed will cause a whole bunch of NoSuchMethodError errors to be spit out. Also changes to the plugin manager method which accounts for any scenarios where the plugin doesn't get enabled (e.g. where the plugin will not be null but won't be enabled because a dependency isn't installed).