Rewrite a large portion of VeinMiner's libraries
- Shift out as much of VeinMiner's code into a new VeinMiner-Common subproject referencing no Bukkit API
- Abstract away the networking to better support packet sending on Fabric
- Drop current client nofication settings (e.g. "Click here to download the client mod!") as they were annoying
- Improve readability and structure of Gradle projects

There are more improvements that need to be done on VeinMiner before 2.0.0 can be considered released. With more features planned, VeinMiner 2.0.0 will be a far more generic and reusable codebase that hopefully paves the way for an upcoming Forge implementation of the client-sided mod. Additionally, this version seeks to reduce the amount of old spaghetti code introduced to VeinMiner over the 6+ years that this project has been maintained. This is an effective rewrite of VeinMiner. All parts of the project will be touched and improved upon.