Send a vein mine request message when target block changes
This can likely be improved. This packet is sent at high frequency if the player just holds the vein mine button and shuffles their mouse around at high speeds, so perhaps a rate limiter on the server may be in order. Also just code wise... it's not pretty. The listeners need their own classes soon.