Support the legacy protocol (temporary)
Because VeinMiner 2.0.0 makes use of a new protocol channel, clients that update to this version of the mod won't be able to join or use the mod on servers that are still running VeinMiner 1.17.10 on the server. This change should still send legacy packets to the server so that even 2.0.0 client mod users can use it on legacy servers.

It should be expected for this legacy support to be removed once 2.0.0 reaches > 50% share on servers (according to bStats).