Remove VeinMinerBlock#getType(), #getState(), and #hasState()
These were methods unused by VeinMiner and were damaging the abstract nature of VeinMinerBlock. The type nor block state needs exposing in VeinMinerBlock because as far as that type is concerned, it should have no need to know about that data. Only the state string is important.

This does keep #getType() in VeinMinerBlockType and #getState() in VeinMinerBlockState, though they are not overridden and cannot be accessed from VeinMinerBlock.